+++ src/components/Table/InvestTable.js
@@ -3,6 +3,7 @@ import LoadingIcon from "../UI/LoadingIcon";
 import UserContext from "../../store/user-context";
 
 import classes from "./Table.module.css";
+import { useHistory } from "react-router";
 
 export const InvestsTitle = (props) => {
   return (
@@ -26,7 +27,7 @@ export const InvestsTitle = (props) => {
 
 export const InvestTile = (props) => {
   return (
-    <div className={classes.tile}>
+    <div className={classes.tile} onClick={props.onClick}>
       <div className={classes.index}>{`${props.index + 1}`}</div>
       <div className={classes.group}>
         <div className={classes.icon}>
@@ -38,7 +39,7 @@ export const InvestTile = (props) => {
         <div className={classes.title}>{props.pool.token0.symbol}</div>
       </div>
       <div className={classes.data}>{`${props.fiat.dollarSign} ${
-        props.pool.tvl || "--"
+        props.pool.tvl.value || "--"
       }`}</div>
       <div className={classes.data}>{`${props.pool.irr || "--"}`} %</div>
       <div className={classes.action}>
@@ -52,6 +53,19 @@ export const InvestTile = (props) => {
 
 const InvestTable = (props) => {
   const userCtx = useContext(UserContext);
+  const history = useHistory();
+  const selectHandler = (option) => {
+    console.log(`option`, option);
+    if (!option.contract) {
+      history.push({
+        pathname: `/import-token/${option.token0.contract}`,
+      });
+    } else {
+      history.push({
+        pathname: `/asset/${option.token0.contract}`,
+      });
+    }
+  };
   return (
     <div className={`${classes.table} ${classes.invest}`}>
       <div className={classes.header}>Invest</div>
@@ -70,6 +84,7 @@ const InvestTable = (props) => {
                 pool={pool}
                 fiat={userCtx.fiat}
                 key={pool.id}
+                onClick={() => selectHandler(pool)}
               />
             ))}
           {props.isLoading && <LoadingIcon />}
