+++ src/components/Table/TokenTable.js
@@ -3,6 +3,7 @@ import LoadingIcon from "../UI/LoadingIcon";
 import UserContext from "../../store/user-context";
 
 import classes from "./Table.module.css";
+import { useHistory } from "react-router";
 
 export const TokensTitle = (props) => {
   return (
@@ -30,7 +31,7 @@ export const TokensTitle = (props) => {
 
 export const TokenTile = (props) => {
   return (
-    <div className={classes.tile}>
+    <div className={classes.tile} onClick={props.onClick}>
       <div className={classes.index}>{`${props.index + 1}`}</div>
       <div className={classes.group}>
         <div className={classes.icon}>
@@ -64,6 +65,20 @@ export const TokenTile = (props) => {
 
 const TokenTable = (props) => {
   const userCtx = useContext(UserContext);
+  const history = useHistory();
+
+  const selectHandler = (option) => {
+    console.log(`option`, option);
+    if (!option.contract) {
+      history.push({
+        pathname: `/import-token/${option.contract}`,
+      });
+    } else {
+      history.push({
+        pathname: `/asset/${option.contract}`,
+      });
+    }
+  };
   return (
     <div className={`${classes.table} ${classes.token}`}>
       <div className={classes.header}>Tokens</div>
@@ -80,6 +95,7 @@ const TokenTable = (props) => {
                 token={token}
                 fiat={userCtx.fiat}
                 key={token.id}
+                onClick={()=>selectHandler(token)}
               />
             ))}
           {props.isLoading && <LoadingIcon />}
