+++ src/constant/constant.js
@@ -1,6 +1,7 @@
 export const uniswapFactory_v2 = `0x5C69bEe701ef814a2B6a3EDD4B1652CB9cc5aA6f`;
 export const uniswapRouter_v2 = `0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D`;
 export const TideBitSwapRouter = `0xDDBCB302A16f27D12Ef1cA491b4791a7b3d67c04`;
+// export const TideBitSwapRouter = `0x3753a62d7654b072aa379866391e4a10000dcc53`;
 export const BinanceSwapRouter = `0x214798a5ca2Fc1cD0d9E4020eCA384406AB67755`;
 export const tidetime = {
   chainId: "0x1f51",
